[DeclChecker] Skip contextualization of initializers associated with property wrapped variables
Fixes a double-contextualization that results in an assert because
properties with property wrappers have the same initializer in two
places - original property (which becomes computed) and backing
storage property - that's where the initializer is copied.

Contextualization of the fully type-checked initializer should happen
as part of backing storage property processing by the declaration
checker.

Resolves: rdar://163562182