[cxx-interop] Use clang lookups for std::vector conformance
Instead of looking up the *imported* Swift type aliases, this patch
shifts the conformance to look up these typedefs from Clang, and *then*
imports them to Swift types to satisfy CxxVector conformance. Doing so
removes the conformance's dependency on eagerly importing such typedefs.

This patch also drops a conformance check that RawIterator conforms to
UnsafeCxxRandomAccessIterator. It shouldn't be necessary, because we are
looking at std::vector, which we assume comes from a conforming stdlib.
Even if it were necessary, it's not clear that this is the right place
and strategy for doing conformance checking. Besides we are fairly
inconsistent about checking other associated types. Let's be optimistic
about conformance for now (-: