[cxx-interop] Use clang lookups for std::map conformance
Instead of looking up the *imported* Swift type aliases, this patch
shifts the conformance to look up these typedefs from Clang, and *then*
imports them to Swift types to satisfy CxxVector conformance. Doing so
removes the conformance's dependency on eagerly importing such typedefs.
There is one remaining dependency on eager imports, which is the
unavailability patching we do for the subscript operator.

This patch also drops a conformance check that the iterators conform to
the our Cxx iterator protocols. It shouldn't be necessary, because we are
looking at std::map, which we assume comes from a conforming stdlib.