[LLDB|CAS] Implement ExplicitCASModuleLoader::addExplicitModulePath()
This virtual function is called when parsing the control block of a Swift module
to add the dependencies to the module loader's explicit Swift module map. For
the compiler this has no practical effect since the explicit Swift module map
should already cover all dependencies. LLDB doesn't have the explicit Swift
module map and depends on this mechanism to discover dependencies.

rdar://166494766