[cxx-interop] Use clang lookups for std::optional conformance
Instead of looking up the Wrapped type from the type of the computed
variable .pointee, this patch shifts the conformance logic to look up
the value_type typedef from Clang, and uses the imported version of that
to to satsify the CxxOptional conformance. Doing so removes the
conformance's dependency on eagerly importing the .pointee method, and
makes the conformance logic consistent with the rest of the CxxStdlib.