[cxx-interop] Always import function return type
There isn't a very good reason _not_ to import the return type when we
already import the argument types, and can occasionally lead to the
return type being misinterpreted as void when our heuristics are wrong.
Even if we're not binding the return value in Swift, we still need to,
say, call its destructor.

rdar://139225705