[ModuleLoader] Remove `llvm_unreachable` in findModuleFilesInDirectory()
In the explicit module loaders, findModuleFilesInDirectory() is not a
supported operation and swift compiler is not expected to call this
function, thus it uses `llvm_unreachable`. However, LLDB and other
clients can configure module loader that allow fallbacks from explicit
module loaders, thus we should function calls to actually return and not
terminating the process. Just let the function return `not_supported`.

rdar://166557713