Sema: apply scoping SuppAssocType scoping rule roots
The generic parameter at the root of a DependentMemberType is
the main thing we care about in terms of the scoping rule.

We have a check for ineffective / invalid inverses in a later
phase, so if the suppression on the type member was wrong then
we should still raise an error about it.