[Overlays] Set `use-external-names` to true in overlay file
`use-external-names` needs to be true in the overlay file, otherwise the
incremental build will not work. If the value is false, the dependency
scanner is reporting virtual path as dependency, thus build system
cannot correctly validate if the dependencies are up-to-date since it
does not know the on disk representation.