LargeTypesReg2Mem: Avoid building an explosion schema just to count the number of registers for very large types
Computing the explosion schema for large types can be expensive.

E.g the following example would spend multiple seconds to compute the
explosion scheme for the array type. It is not worth adding the
complexity to have a special case explosion schema that handles types
like that since exploding a value like that is going to be detrimental.

```
struct Test {
  private var values: [0xfffffff of [UInt32]] =
  InlineArray(repeating: [])
}
```

rdar://165202495