don't share MemberVariables module when testing diagnostics
This header is shared by 3 tests, one of which checks diagnostics. If
that test picks up an already built module from the module cache, it
won't pick up diagnostics in the header. This duplicates the header as
two separate modules rather than specifying an explicit module cache, to
avoid rebuilding CxxStdlib.