SILGen: Fix assert when emitting re-abstraction thunk in edge case involving opaque return type
If we can see the underlying type of an opaque return type, and this
underlying type does not depend on the opaque return type's generic
signature, then we can end up in a situation where the lowered type
of a re-abstraction thunk is not generic, but the formal type is.

This would trigger an assertion failure. Check both the lowered type
and formal type for type parameters to cope with this.

- Fixes https://github.com/swiftlang/swift/issues/86118.