[cxx-interop] [NFC] Make getClangOwningModule a free-standing function
Remove the ClangImporter::Implementation::getClangOwningModule() method,
and replaces all call sites with the static free-standing function it
calls, now exposed in the swift::importer namespace as an internal API.
This makes it easier to use in contexts where an Implementation instance
is not readily available.

Also, clone the documentation to the public ClangImporter API of the
same name.