[cxx-interop] [NFC] Move ClangDirectLookupRequest to ClangLookup.cpp
ClangImporter.cpp is so large that it is getting difficult to navigate.
Start to migrate lookup-related code to a separate file, ClangLookup.cpp.

This first patch moves ClangDirectLookupRequest and related helpers to a new
file, ClangLookup.cpp. This request looks for decls imported from Clang
according to their Swift name, from mappings stored in the SwiftLookupTable
that ClangImporter writes into (its copy of ) the Clang modules it imports.
This request also includes separate handling of lookup in class template
specializations, which are not in the SwiftLookupTable.

At this time of refactoring, lookup works like this. If we are looking in some
struct Foo for a member Bar, we look for _everything_ named "Bar" in the module
that owns Foo, and then filter out everything that isn't a member of Foo
(according to DeclContext). This is inefficient, but correctly accounts for
members that have been "injected" into structs via the SIWFT_NAME, and also
accounts for other name mappings done by the NameImporter.

For the inevitable future Git archaeologist, what follows are the commits whose
blame is destroyed by this refactoring:

    b8e52a7a [cxx-interop] Lazily import members of Clang namespaces and records via requests.
    0ca8dd36 [nfc][cxx-interop] Add three requests `ClangDirectLookupRequest`, `CXXNamespaceMemberLookup`, and `ClangRecordMemberLookup`.
    1089959b [interop] clang name lookup should find declarations in inline namespaces
    c952fc17 [cxx-interop] Fix lookup of class template instantiations
    0eed1a94 [cxx-interop] correctly add and lookup enumerators added to enums that correspond to namespaces