[cxx-interop] [NFC] Move CXXNamespaceMemberLookup to ClangLookup.cpp
ClangImporter.cpp is so large that it is getting difficult to navigate.
Migrate lookup-related code to a separate file, ClangLookup.cpp.

This patch moves CXXNamespaceMemberLookup and its related helper. This request
looks for decls in a namespace, by performing a ClangDirectLookupRequest in each
redeclaration of that namespace. The results are imported to convert them to
swift::ValueDecls.

For the inevitable future Git archaeologist, what follows are the commits whose
blame is destroyed by this refactoring:

    b8e52a7a [cxx-interop] Lazily import members of Clang namespaces and records via requests.
    0ca8dd36 [nfc][cxx-interop] Add three requests `ClangDirectLookupRequest`, `CXXNamespaceMemberLookup`, and `ClangRecordMemberLookup`.
    0c7b1cee [cxx-interop] serialize x-refs for class template specializations
    84a4a8be [Importer] Ensure that we can see macro-expanded declarations in C++ namespaces