Sema: Stop downgrading to private enums cases and destructors with errors
Recent commits changed the behavior downgrading enum cases with errors
to private in `AccessLevelRequest`. It checks `hasInterfaceType` likely
to avoid circular references. This is very prone to side effects. It is
dependent on the timing of the check but the result was cached, making
this behavior unreliable. Let's remove the downgrade to private
behavior.